+++ fs/dev/tf/scenarios/async.fail.ts
@@ -1,0 +1,4 @@
+export const sleep_fail = async () => {
+    await new Promise<void>(resolve => setTimeout(resolve, 10))
+    throw 'async failure'
+}