+++ fs/dev/tf/scenarios/async.pass.ts
@@ -1,0 +1,3 @@
+export const sleep = async () => {
+    await new Promise<void>(resolve => setTimeout(resolve, 10))
+}